Setup any computed default values for scene-wide options
Some of the Camera and Integrator subclasses have default values that must be computed from the Film object. We can only reliably compute these at or after the WorldBegin directive, so that's what we now do.

+++ main.cpp
@@ -544,7 +544,7 @@ namespace minipbrt {
     triCounts.reserve(scene->shapes.size());
     byteCounts.reserve(scene->shapes.size());
 
-    for (uint32_t i = 0, endI = scene->shapes.size(); i < endI; i++) {
+    for (size_t i = 0, endI = scene->shapes.size(); i < endI; i++) {
       if (scene->shapes[i]->type() != ShapeType::TriangleMesh) {
         continue;
       }
